RedHat doesn't let you add scripts during the boot process. For the NetInstalls, this means that the services init script (the one responsible for making sure that services start up, or don't, on bootup) doesn't work correctly on NetInstalls.
The first time you enable a service that was previously disabled, on the next
reboot, the services_init_script will add the various init scripts. However,
because those scripts weren't there when the boot process started, RedHat won't
run them. They will only be run on subsequent reboots as the init script will
be there when the boot process starts.

This changes the services_init_script to be run at both startup, and shutdown.
The startup is mainly applicable to the LiveCD case where services need to be
disabled during the boot process, and the shutdown is mainly applicable to the
NetInstall case, where services need to have been enabled before the boot
process starts.



git-svn-id: https://svn.internet2.edu/svn/perfSONAR-PS/trunk@4507 6dbdcc9b-cb27-0410-b2e0-cf3f5b33df3b